+++ core/src/actions/base.ts
@@ -87,18 +87,20 @@ const actionSourceSpecSchema = createSchema({
   description: dedent`
     By default, the directory where the action is defined is used as the source for the build context.
 
-    You can override this by setting either \`source.path\` to another (POSIX-style) path relative to the action source directory, or \`source.repository\` to get the source from an external repository.
+    You can override the directory that is used for the build context by setting \`source.path\`.
 
-    If using \`source.path\`, you must make sure the target path is in a git repository.
-
-    For \`source.repository\` behavior, please refer to the [Remote Sources guide](${DOCS_BASE_URL}/advanced/using-remote-sources).
-  `,
+    You can use \`source.repository\` to get the source from an external repository. For more information on remote actions, please refer to the [Remote Sources guide](${DOCS_BASE_URL}/advanced/using-remote-sources).`,
   keys: () => ({
     path: joi
       .posixPath()
       .relativeOnly()
       .description(
-        `A relative POSIX-style path to the source directory for this action. You must make sure this path exists and is in a git repository!`
+        dedent`
+          A relative POSIX-style path to the source directory for this action.
+
+          If specified together with \`source.repository\`, the path will be relative to the repository root.
+
+          Otherwise, the path will be relative to the directory containing the Garden configuration file.`
       ),
     repository: joi
       .object()
@@ -109,7 +111,6 @@ const actionSourceSpecSchema = createSchema({
         `When set, Garden will import the action source from this repository, but use this action configuration (and not scan for configs in the separate repository).`
       ),
   }),
-  oxor: [["path", "repository"]],
   meta: { name: "action-source", advanced: true, templateContext: ActionConfigContext },
 })