+++ core/src/vcs/git-repo.ts
@@ -177,7 +177,6 @@ export class GitRepoHandler extends AbstractGitHandler {
 
     const filtered = this.filterPaths({
       files: filesAtPath,
-      log,
       path,
       augmentedIncludes,
       augmentedExcludes,
@@ -191,14 +190,12 @@ export class GitRepoHandler extends AbstractGitHandler {
   }
 
   private filterPaths({
-    log,
     files,
     path,
     augmentedIncludes,
     augmentedExcludes,
     filter,
   }: {
-    log: GetFilesParams["log"]
     files: VcsFile[]
     path: string
     augmentedIncludes: string[]
@@ -214,7 +211,6 @@ export class GitRepoHandler extends AbstractGitHandler {
       // Previously we prepended the module path to the globs
       // but that caused issues with the glob matching on windows due to backslashes
       const relativePath = p.replace(`${path}${sep}`, "")
-      log.silly(() => `Checking if ${relativePath} matches include/exclude globs`)
       return matchPath(relativePath, augmentedIncludes, augmentedExcludes)
     })
   }